Update generic_electrical.dm (#4826)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Allows you to cut the "Wire Tangle" hazard with wirecutters. Adds a
can_be_disabled flag to the hazard that lets you disable it.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Currently the Wire Tangle is completely invincible and players have to
either dig through the walls surrounding it or Ahelp for the hazard to
be destroyed. This will allow players to disable the hazard on their
own.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: Unit I
balance: Wire Tangle hazards can now be cut with wirecutters, rejoice!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->