Resprites the floral shirt+adds two new ones aswell (#4958)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

adds these new shirts to the game, as the old one was really old and we
only had one.

![image](https://github.com/user-attachments/assets/52720bc8-74e3-4757-b7f5-e58e090d3268)

![image](https://github.com/user-attachments/assets/ab859c1a-7861-483a-b0d5-edb1139727ef)


## Why It's Good For The Game

It's summer. We can have some fun

## Changelog

:cl:
add: Resprites the floral shirt and adds two new variants of it

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->