Updates Rabbit Ear Sprites! (#4963)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Does what it says on the tin, updating the bent and the floppy variants!


![image](https://github.com/user-attachments/assets/02e3c3b8-73c1-4874-9ff6-8715d23ee55e)

![image](https://github.com/user-attachments/assets/69ed8cc3-9a61-4c5d-af05-d3b68eb975db)

## Why It's Good For The Game
I had them merged in a rush. Now I feel they're held to a better quality
standard! And also, rabbits are cute!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: PositiveEntropy
imageadd: The sprites for the bent and floppy rabbit ears have been
updated respectively!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->