.300 speedloader fixes and .300 TRAC sprite (#4975)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This fixes the speedloader for the Scout so that it can take specialty
ammo. I'm uncertain if .300 TRAC had a sprite previously, but this PR
should add one now, as the item is currently invisible due to the lack
of a sprite.

## Why It's Good For The Game

This PR brings a few minor fixes

## Changelog

:cl:
fix: fixed .300 speedloaders so that they can take .300 TRAC, and also
gives .300 TRAC sprites
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: Theos <theubernyan@gmail.com>