adds teceti foods to cybersun ships (#4979)
## About The Pull Request

adds the tecetian foods to the cybersun ships

## Changelog

:cl:
add: cybersun ships now have tecetian foods on them
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: thearbiber <135081923+thearbiber@users.noreply.github.com>