Minefield (#2953)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Makes minefields use this cool math I made.
It should have a functionally even distribution 
https://www.desmos.com/calculator/ygix45b89d

small one

![image](https://github.com/user-attachments/assets/76d83e04-55ed-47fb-8bef-4570be6f81fe)
large one

![image](https://github.com/user-attachments/assets/8106981d-0da5-4862-9f17-69f1ae0c987c)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Makes minefields look a bit nicer and lets us set a maximum amount of
mines spawned (There is no minimum so it doesn't freak if it spawns in a
mountain)

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added warning signs to mine fields
refactor: Refactored minefield spawned to be more better c:
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->