Makes Hexacrete have a sane recipe (#5014)
## About The Pull Request

Replaces Phenol in Hexacrete's recipe with Stable Plasma. 

## Why It's Good For The Game

This makes Hexacrete actually worth making, as Phenol is a pain to mix
in bulk to make walls and flooring. Concrete already is used pretty
rarely, and you never see Hexacrete because of the Phenol requirements.

## Changelog

:cl:
balance: Hexacrete now requires Stable Plasma to create.
/:cl: