Fixes build.js so that it properly prepends #define CBS, suppressing compiler warnings (#5033)
@MarkSuckerberg 

I wanted to do more than this but it turns out that _compiler_options.dm
is too late in the dme file order, so I can't move FULL_INIT out of
_basemap.dm without more sweeping changes and that's just out of scope.

So here's the CBS fix (and a tweak on the FULL INIT comment)

Tested on frontier starworks with TGS v6.17.0. This will cleanly prepend

```
#define TGS
#define CBT
```

to the DME before building, so you no longer need to use TGS's
codemodifications menu to add `#define CBT`, the buildchain does it
properly now.

Thanks to @FalloutFalcon for finding the `join` solution. The function
was adding commas without it.

## Changelog

:cl: Anticept, FalloutFalcon 
server: the build chain used in TGS has been fixed so that "#define CBS"
is properly prepended to shiptest.dme, and suppresses compiler warnings
regarding building the dme directly.
/:cl:

Co-authored-by: Mark Suckerberg <mark@suckerberg.gay>