Makes some adjustments to the Harrier. (#5071)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Makes a variety of minor additions and changes to the Harrier-class,
without changing any core design features of it.
Complete list of changes:
<details><summary>Cargo:</summary>
<p>

- Swapped the miner's engineering goggles to just mesons.

- Added a spare jumpsuit for the miner.

</p></details>
<details><summary>Canteen:</summary><p>

- Added card games, dice, plushies, and moved the guitar from
Engineering here.

- Added a bookshelf

- Added universal enzyme to the kitchen for making cheese and other
recipes.

- Added several fruits and vegtables as well as some misc supplies to
the kitchen cabinet to feed the many crewmembers.

</p></details>
<details><summary>Crew Quarters:</summary><p>

- Added pens, crayons, and a paper bin for recreational writing.

- Added an easel and art closet for painting.

- Added spare Assistant clothes and shoes to the laundry cart.

</p></details>
<details><summary>Security Office:</summary><p>

- Replaced the Mini Energy gun with an E-TAR, as well as gave it a spare
cell(the mini-e only had one cell).

</p></details>
<details><summary>Bridge:</summary><p>

- Added an extra weapon power cell to the Captain's locker for the
energy rifle in the gun locker(same story as mini-e).

- Fixed one window grille of the Bridge not being electrified.

- Added a box of spare PDAs to the FO's locker, as well as a PDA
painter.

- Added some decor to the FO's office to make it match the Captain's.

- Added a trans flag next to the fax(VERY IMPORTANT)

- Replaced the coffee vendor with a cheap coffee machine.

- Swapped the placement of the alert console and sensors console.

</p></details>
<details><summary>Cryogenics:</summary><p>

- Added radio closet.

</p></details>
<details><summary>General:</summary><p>

- Changes shuttle window spawners to grille+shuttle windows.

- Adds fitting cleanables throughout the ship.

- Removed role spawn locations.

- Added a object of a filled glass of Bridge Bunny.

- Fixed bug where a flock of birds would descend on the ship and cause
chaos:


![dreamseeker_e9bVFLEsgC](https://github.com/user-attachments/assets/87c4785a-9fed-4108-b88a-3bd5f749bbc3)
</p></details>
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
All the above changes have been done as a result of my extended time
captaining the Harrier in-game, and the issues I have run into. Each
change here I believe will help the ship toward a more enjoyable
experience for players, without invalidating or changing the core
gameplay of the ship.

Cleanables, toys, and art:
The crew of the Harrier is very large, as such the entire crew cannot
participate in a mission the same way the entire crew of a Kilo would,
as there simply is not enough things for them to all take part in a
meaningful fashion, or without being left behind. This is okay though,
as long as staying behind does not mean the players are denied the
ability to interact with the players. The addition of these will
hopefully provide more things for players to do during downtime or if
they need to stay at the ship.

Extra clothes:
Clothes can be destroyed for a number of reasons, and since they aren't
purchasable it's good to have spares. The Harrier already had some, but
I have added the missing ones for consistency.

Extra food:
The Harrier can have up to 13 crewmembers total, and the current pantry
barely provides enough for everyone to be fed out of Cryo, let alone fed
throughout the day(Not even factoring some crewmembers may have diet
restrictions). Other ships with far less crew have been granted massive
pantries, such as the Corona, Ivory, and Nimbus. This PR will bring the
Harrier's stores in line with them more, while still limiting the amount
to give incentive for ingredient collection or purchasing later. I
believe the Harrier should also have a strong kitchen, since it, aside
from the Sunskipper, is the only ship with a dedicated chef.

Weapons:
The Harrier is among the least-equipped ships in the game, despite its
large crew. While I understand that it is a civilian ship, it's very,
very hard to get the ball rolling on the Harrier compared to other
ships, due in part to it's miniscule armory. I added extra power cells
out of consistency, since every gun purchase comes with two magazines, I
thought it would be only fair to give a spare cell to both the energy
rifle and E-TAR. I replaced the energy pistol with the E-TAR because
anyone who's used the energy pistol will tell you that thing does not
count as a real weapon. I didn't want to go overboard with something
like a E-SG or Hades, so I think the E-TAR is a nice boon while not
being too upsetting.

If you would like me to elaborate on any specific change, please comment
below.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Adds a wide variety of changes to the Harrier
code: Added a glass full of Bridge Bunny as a placeable object
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->